requirements.txt
streamlit>=1.37.0
sentence-transformers>=3.0.1
hnswlib>=0.8.0; platform_system != "Windows"
numpy>=1.26.4
pypdf>=4.2.0
python-dotenv>=1.0.1
openai>=1.40.0
google-generativeai>=0.7.2
